+++ projects/qgrid-core/column-list/column.list.sort.js
@@ -9,9 +9,9 @@ function sortIndexFactory(model) {
 	return (columns, scores) => {
 		const { length } = columns;
 		scores = Object.assign({
-			list: column => (column.class === 'data' || column.class === 'cohort') ? 0.1 : 0.3,
+			list: column => (column.category === 'data' || column.category === 'cohort') ? 0.1 : 0.3,
 			index: () => 0.2,
-			view: column => length + ((column.class !== 'data' && column.class !== 'cohort') ? 0.1 : 0.3),
+			view: column => length + ((column.category !== 'data' && column.category !== 'cohort') ? 0.1 : 0.3),
 			template: () => length + 0.4
 		}, scores);